VectorSearch2-0_RAG/requirements.txt
# Vector Search 2.0
google-cloud-vectorsearch>=0.7.0

# Vertex AI (Gemini + Embeddings)
google-cloud-aiplatform>=1.80.0

# GCS (document reading)
google-cloud-storage>=2.19.0,<3.0.0

# PDF extraction
pypdf>=5.3.0

# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0

# Streamlit UI
streamlit>=1.41.0
requests>=2.32.0
